backend/requirements.txt
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
requests==2.33.1
gunicorn==23.0.0
